rmu/pyproject.toml
[project]
name = "rmu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.23.0",
    "dataset>=1.6.2",
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "hf-transfer>=0.1.9",
    "huggingface-hub[hf-xet]>=0.31.4",
    "jupyter>=1.1.1",
    "lab>=8.4",
    "lm-eval>=0.4.2",
    "natsort>=8.4.0",
    "notebook>=7.4.3",
    "optuna>=4.3.0",
    "pandas>=2.2.3",
    "peft==0.5.0",
    "polars>=1.30.0",
    "sentencepiece>=0.2.0",
    "simple-parsing>=0.1.7",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.52.3",
    "wandb>=0.19.11",
]
